+++ src/particles/snow_particle_holder.cxx
@@ -33,7 +33,7 @@ SnowParticleHolder::SnowParticle::SnowParticle (int x, int y, bool colliding_)
   : alive(true),
     colliding(colliding_),
     pos(Vector(x,y)),
-    velocity(Vector(0.0f, 1 + (Math::frand() * 3.5f)))
+    velocity(Vector(0.0f, 1 + ((float)Math::frand() * 3.5f)))
 {
   switch (rand() % 10)
     {
@@ -101,7 +101,7 @@ SnowParticleHolder::update ()
 	  continue;
 	}
 
-      it->velocity.x += (Math::frand() - 0.5f) / 10;
+      it->velocity.x += ((float)Math::frand() - 0.5f) / 10;
 #ifdef CLANLIB_0_6
       if (it->colliding)
         {
